% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{extract_codelist_slug}
\alias{extract_codelist_slug}
\title{Helper function to extract the codelist slug from the OpenCodelists URL}
\usage{
extract_codelist_slug(url)
}
\description{
Helper function to extract the codelist slug from the OpenCodelists URL
}
\keyword{internal}
