% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_semantic_tag}
\alias{extract_semantic_tag}
\title{Extract semantic tag from SNOMED CT description}
\usage{
extract_semantic_tag(string)
}
\arguments{
\item{string}{String, description of SNOMED CT codes}
}
\description{
Add description
}
\examples{
extract_semantic_tag("Blood Pressure (observable entity)")
}
