% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_semantic_tag}
\alias{strip_semantic_tag}
\title{Strip semantic tag from SNOMED CT description}
\usage{
strip_semantic_tag(string)
}
\arguments{
\item{string}{String, description of SNOMED CT codes}
}
\description{
Removes semantic tag from the description
}
\examples{
strip_semantic_tag("Blood Pressure (observable entity)")
}
