% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_semantic_tag}
\alias{extract_semantic_tag}
\title{Extract semantic tag from SNOMED CT description}
\usage{
extract_semantic_tag(string)
}
\arguments{
\item{string}{String, ADD MORE INFO FOR PARAMETER}
}
\description{
Add description
}
\examples{
# Add examples later
}
