% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_semantic_tag}
\alias{strip_semantic_tag}
\title{Strip semantic tag from SNOMED CT description}
\usage{
strip_semantic_tag(string)
}
\arguments{
\item{string}{String, ADD MORE INFO FOR PARAMETER}
}
\description{
Removes semantic tag and returns a  short description
}
\examples{
# Add examples later
}
